git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4547@1433035 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/codecs/BlockTreeTermsReader.java
@@ -276,13 +276,13 @@ public class BlockTreeTermsReader extends FieldsProducer {
    */
   public static class Stats {
     /** How many nodes in the index FST. */
-    public int indexNodeCount;
+    public long indexNodeCount;
 
     /** How many arcs in the index FST. */
-    public int indexArcCount;
+    public long indexArcCount;
 
     /** Byte size of the index. */
-    public int indexNumBytes;
+    public long indexNumBytes;
 
     /** Total number of terms in the field. */
     public long totalTermCount;
